+++ src/Facades/Locales.php
@@ -7,13 +7,13 @@ use Illuminate\Support\Facades\Facade;
 /**
  * Class Locales
  *
- * @method static array get()
- * @method static object current()
- * @method static void set($locale)
+ * @method static \Laraeast\LaravelLocales\Enums\Language[] get()
+ * @method static \Laraeast\LaravelLocales\Enums\Language current()
+ * @method static void set(string|\Laraeast\LaravelLocales\Enums\Language $locale)
  * @method static string getCode()
  * @method static string getName()
  * @method static string getDir()
- * @method static string getFlag()
+ * @method static \Illuminate\Support\HtmlString getSvgFlag(string|int $width = 30, string|int $height = 30)
  * @package Laraeast\LaravelLocales\Facades
  */
 class Locales extends Facade