+++ src/Providers/LocalesServiceProvider.php
@@ -17,10 +17,6 @@ class LocalesServiceProvider extends ServiceProvider
         $this->publishes([
             __DIR__.'/../../config/locales.php' => config_path('locales.php'),
         ], 'locales:config');
-
-        $this->publishes([
-            __DIR__.'/../../flags' => public_path('images/flags'),
-        ], 'locales:flags');
     }
 
     /**