+++ tests/TestCase.php
@@ -2,6 +2,7 @@
 
 namespace Tests;
 
+use Laraeast\LaravelLocales\Enums\Language;
 use Laraeast\LaravelLocales\Facades\Locales;
 use Orchestra\Testbench\TestCase as OrchestraTestCase;
 use Laraeast\LaravelLocales\Providers\LocalesServiceProvider;
@@ -20,18 +21,8 @@ class TestCase extends OrchestraTestCase
         $this->app['config']->set([
             'locales' => [
                 'languages' => [
-                    'en' => [
-                        'code' => 'en',
-                        'name' => 'English',
-                        'dir' => 'ltr',
-                        'flag' => '/images/flags/us.png'
-                    ],
-                    'ar' => [
-                        'code' => 'ar',
-                        'name' => 'العربية',
-                        'dir' => 'rtl',
-                        'flag' => '/images/flags/sa.png'
-                    ],
+                    Language::EN,
+                    Language::AR,
                 ],
             ]
         ]);
@@ -59,5 +50,63 @@ class TestCase extends OrchestraTestCase
         return [
             'Locales' => Locales::class,
         ];
+    }
+
+    /**
+     * Minify html content.
+     *
+     * @param $input
+     * @return string|string[]|null
+     */
+    protected function minifyHtml($input): string|array|null
+    {
+        if (trim($input) === "") {
+            return $input;
+        }
+        // Remove extra white-space(s) between HTML attribute(s)
+        $input = preg_replace_callback('#<([^\/\s<>!]+)(?:\s+([^<>]*?)\s*|\s*)(\/?)>#s', function ($matches) {
+            return '<'.$matches[1].preg_replace('#([^\s=]+)(\=([\'"]?)(.*?)\3)?(\s+|$)#s', ' $1$2',
+                    $matches[2]).$matches[3].'>';
+        }, str_replace("\r", "", $input));
+
+        return preg_replace(
+            [
+                // t = text
+                // o = tag open
+                // c = tag close
+                // Keep important white-space(s) after self-closing HTML tag(s)
+                '#<(img|input)(>| .*?>)#s',
+                // Remove a line break and two or more white-space(s) between tag(s)
+                '#(<!--.*?-->)|(>)(?:\n*|\s{2,})(<)|^\s*|\s*$#s',
+                '#(<!--.*?-->)|(?<!\>)\s+(<\/.*?>)|(<[^\/]*?>)\s+(?!\<)#s',
+                // t+c || o+t
+                '#(<!--.*?-->)|(<[^\/]*?>)\s+(<[^\/]*?>)|(<\/.*?>)\s+(<\/.*?>)#s',
+                // o+o || c+c
+                '#(<!--.*?-->)|(<\/.*?>)\s+(\s)(?!\<)|(?<!\>)\s+(\s)(<[^\/]*?\/?>)|(<[^\/]*?\/?>)\s+(\s)(?!\<)#s',
+                // c+t || t+o || o+t -- separated by long white-space(s)
+                '#(<!--.*?-->)|(<[^\/]*?>)\s+(<\/.*?>)#s',
+                // empty tag
+                '#<(img|input)(>| .*?>)<\/\1>#s',
+                // reset previous fix
+                '#(&nbsp;)&nbsp;(?![<\s])#',
+                // clean up ...
+                '#(?<=\>)(&nbsp;)(?=\<)#',
+                // --ibid
+                // Remove HTML comment(s) except IE comment(s)
+                '#\s*<!--(?!\[if\s).*?-->\s*|(?<!\>)\n+(?=\<[^!])#s',
+            ],
+            [
+                '<$1$2</$1>',
+                '$1$2$3',
+                '$1$2$3',
+                '$1$2$3$4$5',
+                '$1$2$3$4$5$6$7',
+                '$1$2$3',
+                '<$1$2',
+                '$1 ',
+                '$1',
+                "",
+            ],
+            $input);
     }
 }